pyproject.toml
[tool.poetry]
name = "Share-Music"
version = "0.1.0"
description = "This is a bot you can use to share music to all platforms, using song.link's api."
authors = ["tooboredtocode <bored-coder@tooboredtocode.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7"
discord-py-slash-command = "^2.0"
loguru = "^0.5.3"
jishaku = "^2.0"
pillow = "^8.2"
PyYAML = "^5.4"
sentry-sdk = "^1.1.0"


[tool.poetry.dev-dependencies]
taskipy = "^1.8.1"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"